requirements.txt
torch
torchvision
numpy
pillow
opencv-python
scikit-image
matplotlib
pyqt5
python-vlc
